modal: add size props
